requirements.txt
playwright>=1.42.0
pandas>=2.2.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
